Move relevant tests as well
