+++ lib/holiday-stops/src/test/scala/com/gu/holiday_stops/Fixtures.scala
@@ -1,10 +1,9 @@
-package com.gu.holidaystopprocessor
+package com.gu.holiday_stops
 
 import java.time.LocalDate
 
-import com.gu.holiday_stops.{Config, HolidayCreditProduct, HolidayStop, HolidayStopProcessor, Oauth, RatePlan, RatePlanCharge, Subscription, ZuoraConfig}
 import com.gu.salesforce.SalesforceAuthenticate.SFAuthConfig
-import com.gu.salesforce.holiday_stops.SalesforceHolidayStopRequest._
+import com.gu.salesforce.holiday_stops.SalesforceHolidayStopRequest.{HolidayStopRequest, HolidayStopRequestActionedCount, HolidayStopRequestEndDate, HolidayStopRequestStartDate}
 import com.gu.salesforce.holiday_stops.SalesforceHolidayStopRequestsDetail._
 
 object Fixtures {
