+++ lib/holiday-stops/src/test/scala/com/gu/holiday_stops/HolidayCreditSpec.scala
@@ -1,8 +1,7 @@
-package com.gu.holidaystopprocessor
+package com.gu.holiday_stops
 
 import java.time.LocalDate
 
-import com.gu.holiday_stops.{Config, CurrentGuardianWeeklySubscription, HolidayCredit, RatePlan, RatePlanCharge}
 import org.scalacheck.Prop.forAll
 import org.scalacheck._
 import org.scalatest.{EitherValues, OptionValues}
