+++ lib/holiday-stops/src/test/scala/com/gu/holiday_stops/HolidayCreditTest.scala
@@ -1,6 +1,5 @@
-package com.gu.holidaystopprocessor
+package com.gu.holiday_stops
 
-import com.gu.holiday_stops.{CurrentGuardianWeeklySubscription, HolidayCredit, RatePlan}
 import org.scalatest.{EitherValues, FlatSpec, Matchers}
 
 class HolidayCreditTest extends FlatSpec with Matchers with EitherValues {
